+++ cmd/cremengine/engine/api/v1actionsHandler_test.go
@@ -33,7 +33,7 @@ func TestGetValidModelActionsResource_OkResponse(t *testing.T) {
 	// given
 	muxUnderTest := buildMuxUnderTest()
 
-	scenarioTomlText := readTestFileAsText("testdata/ValidTestScenario.toml")
+	scenarioTomlText := readFileAsText("testdata/ValidTestScenario.toml")
 
 	// when
 	postContext := TestContext{
@@ -114,7 +114,7 @@ func TestModelActionsPutRequest_NotAllowedResponse(t *testing.T) {
 func TestModelActionsRequestNotCsv_NotFoundResponse(t *testing.T) {
 	muxUnderTest := buildMuxUnderTest()
 
-	scenarioTomlText := readTestFileAsText("testdata/ValidTestScenario.toml")
+	scenarioTomlText := readFileAsText("testdata/ValidTestScenario.toml")
 
 	// when
 	postContext := TestContext{
@@ -153,7 +153,7 @@ func TestModelActionsRequestNotCsv_NotFoundResponse(t *testing.T) {
 func TestModelActionsRequest_BadCsvContent_BadContentResponse(t *testing.T) {
 	muxUnderTest := buildMuxUnderTest()
 
-	scenarioTomlText := readTestFileAsText("testdata/ValidTestScenario.toml")
+	scenarioTomlText := readFileAsText("testdata/ValidTestScenario.toml")
 
 	// when
 	postContext := TestContext{
@@ -193,7 +193,7 @@ func TestModelActionsRequest_BadCsvCells_BadContentResponse(t *testing.T) {
 	g := NewGomegaWithT(t)
 	muxUnderTest := buildMuxUnderTest()
 
-	scenarioTomlText := readTestFileAsText("testdata/ValidTestScenario.toml")
+	scenarioTomlText := readFileAsText("testdata/ValidTestScenario.toml")
 
 	// when
 	postContext := TestContext{
@@ -213,7 +213,7 @@ func TestModelActionsRequest_BadCsvCells_BadContentResponse(t *testing.T) {
 
 	// when
 
-	invalidRequestBody := readTestFileAsText("testdata/InvalidActiveActions.csv")
+	invalidRequestBody := readFileAsText("testdata/InvalidActiveActions.csv")
 
 	context := TestContext{
 		Name: "POST /model/actions request returns 400 (bad request) response",
@@ -239,7 +239,7 @@ func TestModelActionsRequest_GoodCsvContent_OkResponse(t *testing.T) {
 	g := NewGomegaWithT(t)
 	muxUnderTest := buildMuxUnderTest()
 
-	scenarioTomlText := readTestFileAsText("testdata/ValidTestScenario.toml")
+	scenarioTomlText := readFileAsText("testdata/ValidTestScenario.toml")
 
 	// when
 	scenarioPostContext := TestContext{
@@ -258,7 +258,7 @@ func TestModelActionsRequest_GoodCsvContent_OkResponse(t *testing.T) {
 	verifyResponseStatusCode(muxUnderTest, scenarioPostContext)
 
 	// when
-	validRequestBody := readTestFileAsText("testdata/ValidActiveActions.csv")
+	validRequestBody := readFileAsText("testdata/ValidActiveActions.csv")
 
 	actionsPostContext := TestContext{
 		Name: "POST /model/actions request returns 200 (ok) response",